requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
babel==2.17.0
billiard==4.2.1
boto3==1.40.6
botocore==1.40.6
celery==5.5.3
celery-types==0.23.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.4.0
cryptography==45.0.6
dj-rest-auth==7.0.1
django==4.2.23
django-allauth==65.11.0
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-filter==25.1
django-guardian==3.0.3
django-money==3.5.4
django-storages==1.14.6
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
flower==2.0.1
gunicorn==22.0.0
humanize==4.13.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pip==25.2
pluggy==1.6.0
prometheus-client==0.22.1
prompt-toolkit==3.0.51
psycopg2-binary==2.9.9
py-moneyed==3.0
pycparser==2.22
pyjwt==2.10.1
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
ruff==0.12.8
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.4.0
tornado==6.5.2
types-pyyaml==6.0.12.20250809
typing-extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uv==0.8.8
vine==5.1.0
wait-for-it==2.2.2
wcwidth==0.2.13
wheel==0.45.1
